requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
astunparse==1.6.3
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
h5py==3.9.0
idna==3.4
imageio==2.31.1
imgaug==0.4.0
itsdangerous==2.1.2
jax==0.4.13
Jinja2==3.1.2
joblib==1.3.1
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
mtcnn==0.1.1
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
SQLAlchemy==1.4.39
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
typing_extensions==4.7.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.2.1
wrapt==1.14.1
